Refactor unit tests
Add base class for unit tests, that can setup any application. The base
class creates a workflow, branch provider and os service, and generates
a folder tree.
